refactor(cmd): extract environment detection to environment.go
